Implemented the ASCII view in full. Both calculators also support decimal entries.

+++ Calculator-CS2800-2/src/main/java/uk/ac/rhul/cs2800/StandardCalc.java
@@ -55,7 +55,7 @@ public class StandardCalc implements Calculator {
    */
   private static boolean isNumeric(String c) {
     try {
-      Integer.parseInt(c);
+      Float.parseFloat(c);
       return true;
     } catch (Exception e) {
       return false;
